requirements.txt
# Django and REST framework
Django
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==2.3.0

# Database
djongo
pymongo==3.12.3

# Real-time capabilities
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0
websockets==12.0

# AI/ML
tensorflow==2.14.0
scikit-learn==1.3.2

# Utils
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.2
numpy==1.26.1

# Development
black==23.10.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.5.2
django-model-utils==4.3.1
